Removed all protected members from road (made them private). Fixed some compile time warnings
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3913 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/worker.cc
@@ -1522,7 +1522,7 @@ void Worker::program_update(Game & game, State & state)
 		WorkerProgram const * program =
 			static_cast<WorkerProgram const *>(state.program);
 
-		if (state.ivar1 >= program->get_size())
+		if (static_cast<uint32_t>(state.ivar1) >= program->get_size())
 			return pop_task(game);
 
 		Action const & action = *program->get_action(state.ivar1);
